Fix when not used with l10n_fr and refactor shareholder register tests
Put all common test code in... common instead of duplicating it, adding a proper
way to make the tests pass in a single-module DB as well as a DB with l10n_fr (or
any other l10n module that adds a specific chart of account).
