+++ commown_shareholder_register/tests/test_shareholder_register.py
@@ -5,123 +5,13 @@ from io import BytesIO
 from odf import opendocument
 from odf.table import Table, TableCell, TableRow
 
-from odoo.tests import TransactionCase
+from .common import TestShareholderRegisterTC
 
 
-class TestShareholderRegister(TransactionCase):
-    @classmethod
-    def _create_account(cls, code, name):
-        account = cls.env["account.account"].create(
-            {
-                "name": name,
-                "code": code,
-                "account_type": "income_other",
-            }
-        )
-        return account
-
-    @classmethod
-    def _create_college(cls, name, rank):
-        college = cls.env["commown_shareholder_register.college"].create(
-            {
-                "name": name,
-                "rank": rank,
-            }
-        )
-        return college
-
-    @classmethod
-    def _create_category(cls, name, account, college, min_share_number):
-        category = cls.env["commown_shareholder_register.category"].create(
-            {
-                "name": name,
-                "account_id": account.id,
-                "college_id": college.id,
-                "min_share_number": min_share_number,
-            }
-        )
-        return category
-
-    @classmethod
-    def _add_shares(cls, partner, account, date_tuple, amount):
-        journal = (
-            cls.env["account.journal"]
-            .search([("type", "=", "general")], limit=1)
-            .ensure_one()
-        )
-        move = cls.env["account.move"].create(
-            {
-                "name": "Test Account Move",
-                "journal_id": journal.id,
-                "date": date(*date_tuple),
-            }
-        )
-        if amount < 0:
-            attr2, attr1 = "credit", "debit"
-        else:
-            attr1, attr2 = "credit", "debit"
-
-        cls.account_move_lines |= cls.env["account.move.line"].create(
-            [
-                {
-                    "move_id": move.id,
-                    "account_id": account.id,
-                    "partner_id": partner.id,
-                    "date": date(*date_tuple),
-                    attr1: abs(amount),
-                },
-                {
-                    "move_id": move.id,
-                    "account_id": cls.account_balancing.id,
-                    "partner_id": partner.id,
-                    "date": date(*date_tuple),
-                    attr2: abs(amount),
-                },
-            ]
-        )
-
+class TestShareholderRegister(TestShareholderRegisterTC):
     @classmethod
     def setUpClass(cls):
-
-        super(TestShareholderRegister, cls).setUpClass()
-        cls.env.company = cls.env.ref("l10n_fr.demo_company_fr")
-        cls.env.user.company_id = cls.env.company.id
-        cls.env.company.nominal_share_amount = 20
-
-        cls.partner_1 = cls.env["res.partner"].create({"name": "Partner 1"})
-        cls.partner_2 = cls.env["res.partner"].create({"name": "Partner 2"})
-        cls.partner_3 = cls.env["res.partner"].create({"name": "Partner 3"})
-
-        cls.account_porteur = cls._create_account("10134000", "Porteur de Projet")
-        cls.account_soutient = cls._create_account("10136000", "Account Soutient")
-        cls.account_beneficiaire = cls._create_account(
-            "10135000", "Account Bénéficiaire"
-        )
-        cls.account_balancing = cls._create_account("XXXXXXXX", "Balancing journal")
-
-        cls.college_A = cls._create_college("A", 80)
-        cls.college_B = cls._create_college("B", 30)
-        cls.college_D = cls._create_college("D", 40)
-
-        cls.cat_porteur = cls._create_category(
-            "Porteur",
-            cls.account_porteur,
-            cls.college_A,
-            100,
-        )
-        cls.cat_soutient = cls._create_category(
-            "Soutient",
-            cls.account_soutient,
-            cls.college_D,
-            5,
-        )
-        cls.cat_beneficiaire = cls._create_category(
-            "Beneficiare",
-            cls.account_beneficiaire,
-            cls.college_B,
-            1,
-        )
-
+        super().setUpClass()
         cls.account_move_lines = cls.env["account.move.line"]
         cls._add_shares(cls.partner_1, cls.account_porteur, (2018, 3, 12), 2000)
         cls._add_shares(cls.partner_1, cls.account_porteur, (2018, 8, 12), -2000)
